MB-36296 related - support for collections in Analytics workbench
There are several places in the workbenchs where the user needs to
select a bucket, scope, and collection:

- The document editor, for selecting the current domain for querying
- The import UI, for selecting the import destination
- The analytics dataset UI, for selecting the dataset source

This change adds an Angular 8 component which provides a menu for
selecting bucket, scope, and collection. It also makes the component
available to AngularJS through downgradeCompanent.

The change also updates the doc editor and import code to use the new
component.

Change-Id: Ib5c03050c349fd490f15acd7c332b93f10934016
Reviewed-on: http://review.couchbase.org/c/query-ui/+/142626
Tested-by: Eben Haber <eben@couchbase.com>
Reviewed-by: Eben Haber <eben@couchbase.com>