MB-39400 - fix column sorting in the document editor
It turned out that column sorting had never been implemented in the
new Angular 8 version of the document editor. Not only that, but
horizonal scroll syncing between the header row and the data rows had
also not been added. These proved to be a bit tricky to implement, as
the doc editor uses a ViewChild for the dynamically created and
compiled inner table.

Change-Id: I12eabbddeab1e6c22ccb9069b105e4975088bcb4
Reviewed-on: http://review.couchbase.org/c/query-ui/+/143376
Tested-by: Eben Haber <eben@couchbase.com>
Reviewed-by: Eben Haber <eben@couchbase.com>