MB-41877 - Hide extra horizontal scrollbar in Firefox in data table
The datatable uses a separate <div> for the columns headers, to enable
them to always be visible. During horizontal scrolling, we don't want
that div to have scrollbars, since the div for the main table will
also show a scrollbar. Firefox's css to hide scrollbars has changed
over the years, update to the latest way: scrollbar-width: none.

Change-Id: Ia4a6dde68001d1a7b2e82234d77224164fc5c507
Reviewed-on: http://review.couchbase.org/c/query-ui/+/143554
Tested-by: Eben Haber <eben@couchbase.com>
Reviewed-by: Eben Haber <eben@couchbase.com>