MB-39900 - data-table sort when field names start with spaces
The bug noted a problem where the data table could not sort fields
whose names started with spaces. The table is composed of generated
HTML, based on the data. The sort function used the text of the HTML
in the header, but that trimmed whitespace. The solution is to put the
column name in an html attribute called "data-colname". This worked in
general, but testing revealed issues with column names containing
single and double quotes, in both the data table (read-only) and table
editor. The solution for quotes is to sanitize the keys. (Before the
data values were sanitized, but not the keys.)

Change-Id: I2e3257cd7c2797618f57e21b422cbbbe0782a931
Reviewed-on: http://review.couchbase.org/c/query-ui/+/143732
Tested-by: Eben Haber <eben@couchbase.com>
Reviewed-by: Eben Haber <eben@couchbase.com>