MB-40200 - Fix issue with tabular view of arrays of objects holding arrays.
Query results are always arrays of items. When those items are things
like strings, boolean, numbers, or arrays, the height of each item was
computed correctly. However, when the result was an array of objects,
where each object had a single field whose value was an array, the
heights were not correctly computed. E.g.,

[
{"schedule": [1,2,3]},
{"schedule": [4,5,6]},
{"schedule": [7,8,9]}
]

Change-Id: I1668f6f599dd1e1a7a281707c50d25c43ebe57c5
Reviewed-on: http://review.couchbase.org/c/query-ui/+/143824
Tested-by: Eben Haber <eben@couchbase.com>
Reviewed-by: Eben Haber <eben@couchbase.com>