Fix a resource leak in the doc editor slowing performance
The doc editor had a variety of problems leading to resource
leaks/slow performance. Changed ngIf to [hidden] to avoid
destroying/recreating the inner component every time. Also made sure
to unsubscribe from the Observable when the inner component is
destroyed. Finally, changed to ChangeDetectionStrategy.OnPush, and use
Observable to push the changes.

Change-Id: Ibe356691300f196be52f9d055864574b84d30294
Reviewed-on: http://review.couchbase.org/c/query-ui/+/144182
Tested-by: Eben Haber <eben@couchbase.com>
Reviewed-by: Eben Haber <eben@couchbase.com>