MB-43957 - problem cancelling query from GUI workbench
The problem in this case was that the query was crashing the query
engine, thus the GUI was never getting a reply to the REST request to
run the query. This change marks the query as cancelled and no longer
running if the cancel request succeeded, even if the original REST
request has not returned.

Change-Id: I18d0ed6f90ba336f1ce854d3af65d4abd9d90343
Reviewed-on: http://review.couchbase.org/c/query-ui/+/145213
Tested-by: Eben Haber <eben@couchbase.com>
Reviewed-by: Eben Haber <eben@couchbase.com>