Fix an exception in the doc editor when trying to read files > 5MB.
In 6.5 we put up a notice dialog to warn the user that loading the
file could take some time. Before now, there was no notice dialog in
7.0, leading to an execption in this case.

This change adds the notice dialog to the QwDialogService, shows the
notice on opening a large file, and removes the notice once the file
is completed.

Change-Id: I0835ceed39313c96910264935a30130e88f9a38c
Reviewed-on: http://review.couchbase.org/c/query-ui/+/145229
Tested-by: Eben Haber <eben@couchbase.com>
Reviewed-by: Eben Haber <eben@couchbase.com>