MB-45423 - QWB query failure with queries longer than 60 seconds
The problem here was that the custom header ns-server-proxy-timeout
was not being included in the HTTP request. It was being set in the
code, but not propigated to the eventual request. The problem was as
follows:

When moving to Angular 8, I started using the Angular class
HttpHeaders. There are two ways to set the headers for HttpHeaders:

1) a constructor to which you can pass an object containing the headers, or
2) a 'set' method to set the headers one at a time.

I was using option #1, but it turns out that there is a bug/feature in
which this constructor only sets headers with string values, and
ns-server-proxy-timeout has a number value. The set method works just
fine for both string and number values, so this change moves to using #2.

Change-Id: I2c2a86bf14a98e0ba2ed240ad2b8558af56e634d
Reviewed-on: http://review.couchbase.org/c/query-ui/+/150768
Tested-by: Eben Haber <eben@couchbase.com>
Reviewed-by: Eben Haber <eben@couchbase.com>