MB-45739 - Documents view gets bogged down with large/complex documents
The "Field Editing" mode of the Documents view is a nice feature, but
the browser can get bogged down when trying to render thousands of
INPUT fields. Change the table editor directive to compute the total
number of fields in the data set, and in each document, and to disable
field editing mode if rendering the fields will take too long.

Change-Id: I1674649cfb064eca0e8d7c4b384acd318a8fb694
Reviewed-on: http://review.couchbase.org/c/query-ui/+/151652
Tested-by: Eben Haber <eben@couchbase.com>
Reviewed-by: Eben Haber <eben@couchbase.com>