Fix an exception in the Documents page thrown when there are no buckets.
There was an *ngIf on the mn-element-cargo cargo element (which puts
the "Create Document" button at the top of the page). The *ngIf was
hiding the button when there was no selected bucket. Apparently
mn-element-cargo didn't like it when the element wasn't there. By
switching to [hidden], the element is still there but not visible,
which keeps mn-element-cargo happy.

Change-Id: Ib5337b14496d8d97343e27fa8b28c24757163a91
Reviewed-on: http://review.couchbase.org/c/query-ui/+/151994
Tested-by: Eben Haber <eben@couchbase.com>
Reviewed-by: Eben Haber <eben@couchbase.com>