MB-42043 - Need a better N1QL parser on the client side
This bug was caused by the client-side N1QL parser being out-of-date
with respect to the parser in the query engine. Up to now the
client-side, javascript n1ql parser was created by hand using Jison
and a rough copy of the golang goyacc n1ql parser. Maintaining this
proved to be labor intensive with all the n1ql changes with each
release. This change helps automate the process.

First, I decided to use the cross-language parser generator Antlr.

Second, I created a grammar in Antlr for parsing goyacc files,
rewriting the rules as Antlr. This generates n1ql.g4, an Antlr
definition of the N1QL grammar. This process needs only be run
occassionally, perhaps one or two times per release.

Third, I created a parser "listener", which walks the parse tree to
generate the information needed on the client (basically the path
information for fields used in n1ql queries and expressions).

I added Makefiles to permit the parser generation process to be run as
needed, but in general the version of the parser that is checked in
should be fine.

Change-Id: I35552b07349c76ea150160863d89fea57af8a81d
Reviewed-on: http://review.couchbase.org/c/query-ui/+/152511
Tested-by: Eben Haber <eben@couchbase.com>
Reviewed-by: Eben Haber <eben@couchbase.com>