Upgrade Query Workbench to Angular 8
In release 7.0, the query workbench was a hybrid: the document editor
and import components were Angular 8, while the query workbench and
query monitoring were still in AngularJS. This change upgrades
everything to Angular 8, and removes the obsolete code.

There are companion changes in cbas-ui and fts as well, since both
those plug-ins used services or components from query-ui.

Change-Id: I09397f6c78be081e8d56d2cca4f9b07930999444
Reviewed-on: http://review.couchbase.org/c/query-ui/+/157458
Tested-by: Eben Haber <eben@couchbase.com>
Reviewed-by: Eben Haber <eben@couchbase.com>