MB-47371 - Problem with documents page bucket selector in cluster_run
The bucket/scope/collection selector used in the Documents page is
also used by analytics in selecting buckets/collections for local and
remote datasets (a.k.a. "analytics collections"). In the case of
analytics, the remote cluster may be 7.0 or pre-7.0, and thus may or
may not support collections. The code to test remote versions for
analytics was also getting tripped in the case of cluster_run on the
master branch, where the version is "0.0.0-0000-enterprise", and 0 is
less than 7. The fix is to make sure that to only check the version of
remote clusters, not local ones. And if the remote cluster is
cluster_run, well, we're just out of luck.

With or without this fix, the official builds should never run into
this problem, since they all have a real version number.

Change-Id: I6745f2aedbe1933c77ff1588e863b17d189aa2ec
Reviewed-on: http://review.couchbase.org/c/query-ui/+/157550
Tested-by: Eben Haber <eben@couchbase.com>
Reviewed-by: Eben Haber <eben@couchbase.com>