Support for MB-47574, and also FTS in Angular.
The QW dialog service already can show a doc-editing dialog box. This
change adds a method to retrieve the document first, by docId, then
showing the dialog box with the document JSON. This is necessary for
FTS (which currently uses the soon-to-be-deprecated AngularJS
qw_doc_editor_service), and also the Views UI (MB-47574).

Change-Id: Ic724c3f88fb6fbc1c914c13a0d56e237642dd838
Reviewed-on: http://review.couchbase.org/c/query-ui/+/158319
Tested-by: Eben Haber <eben@couchbase.com>
Reviewed-by: Raluca Lupu <raluca.lupu@couchbase.com>