MB-47795 - date problem with charts
There are two related date problems here. One is that the code accepts
a limited number of date formats. The second was that it was using a
regex to recognize the presence of a date in a string, but the d3 Time
Parser was getting hung up if there were extra characters (such as a
time zone specifier).

To address the problem I have added a few additional recognized
formats, and made sure that the d3TimeParse is applied only to the
part of the string that matched the time format.

Change-Id: I16124ba84c0ec59f270f7d6ef20be2a6368b0e42
Reviewed-on: http://review.couchbase.org/c/query-ui/+/159160
Tested-by: Eben Haber <eben@couchbase.com>
Reviewed-by: Eben Haber <eben@couchbase.com>