MB-47755 - make chart visible after navigation
Bug: run a query, create a chart, navigate to Documents and
back. Chart is gone, though tabular and JSON views still show the
query result.

The query workbench in AngularJS, while the chart component is Angular
8. The best way for the workbench to inform the chart component about
changes is through an Observable, which it does at the end of each
query execution. However, this didn't account for navagition to other
parts of the UI, since the Component is recreated each time the user
navigates there. Add a line to the controller "activate" function to
update the Observable with any available data when the Component is
created.

Also add some error checking to avoid errors in the browser console.

Change-Id: Ifd8e030103199807ea51144c87a95e74239296e3
Reviewed-on: http://review.couchbase.org/c/query-ui/+/159166
Tested-by: Eben Haber <eben@couchbase.com>
Reviewed-by: Eben Haber <eben@couchbase.com>