MB-47832 - better sync chart with other results
The bug noted that issueing an erroneous query did not cause the
current chart to be cleared. The problem was the error state caused
'createChart' to exit before clearing the current chart. The chart
clearing code has been moved before the error checks.

Also make sure that the chart contents are updated when the user goes
back and forth in the history.

Change-Id: I7b087891696aca237eee4499bf3d5a8360069a66
Reviewed-on: http://review.couchbase.org/c/query-ui/+/159175
Tested-by: Eben Haber <eben@couchbase.com>
Reviewed-by: Eben Haber <eben@couchbase.com>