MB-47931 - Y-axis problem with date fields
The problem here was that dates should not be permitted for the
y-axis. The fix updated the type-checking routine to prevent dates on
the y-axis.

Also, slightly reduce the opacity for x-y chart points, so that
multiple points in the same location are more obvious.

Finally, address a problem with very large data sets where an error
message was shown inappropriately for the chart and table views.

Change-Id: I76397cf0d7ae1e6bc0db33f28e57d0e6d834fb72
Reviewed-on: http://review.couchbase.org/c/query-ui/+/159945
Tested-by: Eben Haber <eben@couchbase.com>
Reviewed-by: Eben Haber <eben@couchbase.com>