Fix incorrect "data too big" error message.
The query workbench issues a "data too big" message when the query
result data is too large to be efficiently displayed in the
browser. The message is contingent on the result view, since the table
viewer can display about 10x the data of the JSON viewer. We used to
have a tree viewer, but when it was replaced by the chart, the "data
too big" logic for the tree viewer remained in place. This change
removes the "data too big" logic left over from the tree viewer.

Change-Id: I7127ec412dde60850d6fc1f581c8cf8373c5b310
Reviewed-on: https://review.couchbase.org/c/query-ui/+/166872
Tested-by: Eben Haber <eben@couchbase.com>
Reviewed-by: Eben Haber <eben@couchbase.com>