MB-47263 - Partial fix for 'field not found' issue
When running user queries, the query workbench compares the fields
used in the query with those found in the inferred schema (if any). If
there is a schema but the field is not found, that could be a typo, so
a warning message is shown.

This code was not handling collections properly, as well as some
recent changes to the query plan, so this change allows the warning
message to once appear when appropriate.

Now it will be possible to work on the actual problem in MB-47263.

Change-Id: I11e28391437e6a4dac5128abc4c67a1946d3f015
Reviewed-on: https://review.couchbase.org/c/query-ui/+/166879
Tested-by: Eben Haber <eben@couchbase.com>
Reviewed-by: Eben Haber <eben@couchbase.com>