MB-47263 - spurious warning messages when search_meta in query
The query workbench has a feature for detecting typos by comparing all
query expressions mentioned in the query plan with the fields found in
the inferred schema.

When 'search_score' or 'search_meta' are used in queries, the query
plan includes a reference to the non-existent field 'out', which could
cause a warning message about a possible typo. This change ignores the
query plan expressions related to search_meta and search_score.

Change-Id: Ie3da7aa9319d800a576fce7bffbb902a07a6a16b
Reviewed-on: https://review.couchbase.org/c/query-ui/+/167077
Tested-by: Eben Haber <eben@couchbase.com>
Reviewed-by: Eben Haber <eben@couchbase.com>