MB-50005 - don't allow add function/library if no permissions
The appropriate permissions behavior is a bit more complex than
described in the bug. There are permissions for managing external
function libraries, which are required to even query the list of
libraries. If those permissions are not found, we can't show any
libraries, and have to instead show a permissions message.

For functions, it's a bit different. The user can see the list of
functions if they are able to read the system catalogs, even if they
don't have permission to edit them. Thus we need three different
behaviors:

- no system catalog perms -> no list of functions
- system catalogs, but no manage function perms -> read only functions
- system catalog + manage functions -> full read/write functions

Change-Id: Ice52c2f9a453a1951befeddf9f8f778b4146f3a9
Reviewed-on: https://review.couchbase.org/c/query-ui/+/167339
Tested-by: Eben Haber <eben@couchbase.com>
Reviewed-by: Eben Haber <eben@couchbase.com>