MB-48160 - handle bar charts with too many bars for the window
When d3 attempts to create a bar chart with more bars than available
pixels in the window, the chart appears empty. This change detects
this case for bar and grouped bar chart, and instead shows a message
indicated that the user needs to either enlarge the window or select
less data.

Change-Id: Iaab21251b970bcfdb0899a247efd668b047db9ed
Reviewed-on: https://review.couchbase.org/c/query-ui/+/167347
Tested-by: Eben Haber <eben@couchbase.com>
Reviewed-by: Eben Haber <eben@couchbase.com>