MB-50532 - CSS problems with Analytics Workbench
Due to a recent change (77918d9) the CSS used by various query
Components was included in a new location. Unfortunately, some of
those components were also used in the Analytics workbench, and the
CSS was not included there.

This change makes the Components refer to the CSS they need directly,
so regardless of who uses them the CSS will be available.

We also removed the obsolete datatable Directive, since Angular
Directives are not permitted to specify their own CSS. The datatable
Component works equivalently.

Change-Id: I8577dbe5c61df98bdc34ce51faa8662aa82007b9
Reviewed-on: https://review.couchbase.org/c/query-ui/+/169361
Tested-by: Eben Haber <eben@couchbase.com>
Reviewed-by: Hussain Towaileb <Hussain.Towaileb@Couchbase.com>
Reviewed-by: Eben Haber <eben@couchbase.com>