MB-50324 - problems with import UI and quoted special characters
The UI for import was applying JSON.parse to the data loaded from a
file. This worked incorrectly with quoted special characters, such as
\n. While '\n' is a valid way of expressing a newline in JSON, it is
also the way to express a newline in a javascript string. Thus, inside
javascript code, we need to double the backslashes for JSON.parse to
work. The first backslash gets interpreted by javascript itself, and
the second by JSON.parse. It makes my head hurt just thinking about
it, but it works.

Here is an example of a .json file (in lines format) with various
special charaters:

{"type":"name", "val":"hello\\world\n"}
{"type":"cr", "val":"My cr is: \r"}
{"type":"nl", "val":"My nl is: \n"}
{"type":"tab", "val":"My tab is: \tafter tab"}
{"type":"bs", "val":"My bs is: \bafter bs"}
{"type":"reverseSolidus", "val":"My rs is: \\"}
{"type":"solidus", "val":"My rs is: \/"}
{"type":"bare quote", "val":"My quote is: \"after quote"}

Before the fix, the import UI would report a parse error. After the
fix, all the documents are correctly parsed and imported.

Change-Id: I499fc8309621a732117ae297b6c8a39a64cb5e95
Reviewed-on: https://review.couchbase.org/c/query-ui/+/169453
Tested-by: Eben Haber <eben@couchbase.com>
Reviewed-by: Eben Haber <eben@couchbase.com>