MB-50506 - inline UDF expression issues
In many cases it is necessary to surround the inline query expression
with parentheses to avoid a N1QL syntax error. However the code was
adding parentheses whether or not they were already present. This
change fixes that, parens should only be added if not already present.

Also make sure to encode UDF library names when creating or deleting
libraries, to handle cases where the names have weird characters that
might mess up the URL.

Change-Id: I52f830fd2b7231a10de22b6bda41e36ab5d34c8b
Reviewed-on: https://review.couchbase.org/c/query-ui/+/169853
Tested-by: Eben Haber <eben@couchbase.com>
Reviewed-by: Eben Haber <eben@couchbase.com>