MB-50798 - UDF display issue due to API change
MB-50462 - Permit scoped UDF Libraries

Both issues are part of the same underlying problem: the ability for
the UI to support UDF libraries that are tied to a specific
bucket/scope.

The first issue was caused by an API change where the UDF libraries
are now returned with bucket and scope names, if any.

The second issue is a completion of the UDF UI feature for neo. The
original implementation in the UI was limited to global libraries. The
bulk of the changes here allow libraries to be created within a
specific bucket/scope (and, of course, for functions to be defined in
terms of scoped as well as global libraries).

It should be noted that, according to Kamini, javascript functions are
limited to using libraries in the same scope. E.g., global functions
can only be based on global libraries, and a function in
`travel-sample`.`inventory` can only use libraries defined in that
same bucket/scope.

Change-Id: Ia4050f963f258f6c1111302ad24d29bdc633c189
Reviewed-on: https://review.couchbase.org/c/query-ui/+/170198
Tested-by: Eben Haber <eben@couchbase.com>
Reviewed-by: Eben Haber <eben@couchbase.com>