Remove unncessary provider entries left over from Angular upgrade
When the query workbench was still in AngularJS, it was necessary to
list all the Angular services used by it in the list of providers in
main.js. Otherwise, a "Provider not Found" error would appear when
reloading the browser when viewing the query workbench. Since the
query workbench has now been upgraded to Angular 8, this list is no
longer needed.

Change-Id: I4515d53a68caca159e745cbbf7a1969271d785ee
Reviewed-on: https://review.couchbase.org/c/query-ui/+/170614
Tested-by: Eben Haber <eben@couchbase.com>
Reviewed-by: Eben Haber <eben@couchbase.com>