MB-50952 - Import UI not flagging error with empty field names
Two problems were reported in this issue:

- the import UI was accepting data with blank field names, while
  cbimport would flag that as an error.

- cbimport treated empty values in CSV files as empty strings, while
  the import UI marked those fields as null.

This change flags an error if any field names are empty, and converts
empty CSV values to the empty string.

There may be other differences between cbimport and the import UI,
given that the two are fundamentally different. This change brings
them closer together, however.

Change-Id: I906acfb80aa0797011a3f341de6f19972879dbe0
Reviewed-on: https://review.couchbase.org/c/query-ui/+/170931
Tested-by: Eben Haber <eben@couchbase.com>
Reviewed-by: Eben Haber <eben@couchbase.com>