MB-51073 - Separate QwQueryService into two services.
The QwQueryService started as a simple service for submitting queries,
but it quickly grew to be the persistent store for the user's query
history, query options, and various metadata (such as inferred
schemas). This proved to be a problem for the Index UI, which used the
QwQueryService for dropping indexes.

This change separates the functionality into two services:
QwQueryService is the bare-bones code that sends n1ql queries to the
server. QwQueryWorkbenchService contains all the extra functionality
for supporting the UI workbench: query history, metadata, etc.

The above work was necessary for this change: Remove unncessary
provider entries left over from Angular upgrade

When the query workbench was still in AngularJS, it was necessary to
list all the Angular services used by it in the list of providers in
main.js. Otherwise, a "Provider not Found" error would appear when
reloading the browser when viewing the query workbench. Since the
query workbench has now been upgraded to Angular 8, this list is no
longer needed.

Change-Id: I1876401b1c4f7c030c9000fd3585477562a49649
Reviewed-on: https://review.couchbase.org/c/query-ui/+/171051
Tested-by: Eben Haber <eben@couchbase.com>
Reviewed-by: Eben Haber <eben@couchbase.com>