MB-51176 - Errors when checking for read permission in doc editor
The tabular doc editor has buttons for edit, copy, and delete that
must only be enabled if the user has write permission. The previous
checks sometimes caused a null exception.

Also, for users restricted to a single scope or collection, we need
to check cluster.collection[bucket:.:.] instead of cluster.bucket.
(Related to MB-51147).

Change-Id: I10cc685f229227cbcbc3dbd6d25ae8fed187f953
Reviewed-on: https://review.couchbase.org/c/query-ui/+/171361
Tested-by: Eben Haber <eben@couchbase.com>
Reviewed-by: Eben Haber <eben@couchbase.com>