MB-51222 - Permissions issues in the query workbench
Up to now the UI was checking cluster.collection[].data.docs!write to
see if document editing and importing is permitted. It turns out that
"write" is not used on the server. Instead we need to use:

data.docs!upsert - allows creating docs & editing existing docs
data.docs!delete - deleting documents

(There is also a data.docs!insert permission, but it doesn't seem to
be checked by the docs REST API when creating new documents.)

Given these permissions, and data.docs!read, add a filtered list of
buckets to the qw.metadata.service so we can show only those buckets
for which the user has some level of access. /pools/default lists all
buckets even if no access is permitted.

This change also improves the behavior of MB-51191. The namespace menu
now shows only those buckets and scopes for which the user has some
permissions. However if a user has read access to a scope, but not
udf.manage permissions, the menu will still include that scope. In
addition, a user with Query System Catalogs role can see all buckets
even if they don't have access to them.

Change-Id: I8a0127af99156470f5f9d6f7799deea3a9c868ef
Reviewed-on: https://review.couchbase.org/c/query-ui/+/171601
Tested-by: Eben Haber <eben@couchbase.com>
Reviewed-by: Eben Haber <eben@couchbase.com>