MB-51256 - layout issues with datatable in query workbench
There were two problems here:

1) In Angular the datatable component needs to be inside a div with
position: absolute, a specified width, and a fixed height. This wasn't
true in AngularJS (as can be seen in the Analytics UI), but we added a
style to the css to cover this for the query workbench.

2) There was an off-by-one error in the datatable when computing the
height for arrays. This caused the vertical overlap in the results
reported in the issue. In fact, with data containing top-level arrays,
it was off by much more. Added a short code block to sum heights
properly for arrays.

Change-Id: I40603f99f8261fd3a7e8ac0865f91c88713b22fc
Reviewed-on: https://review.couchbase.org/c/query-ui/+/171700
Tested-by: Eben Haber <eben@couchbase.com>
Reviewed-by: Eben Haber <eben@couchbase.com>