MB-51375 - problem with scoped libraries and functions
Problem 1: when editing a scoped library, the dialog box was not
correctly showing the library's scope, instead showing "global". Some
context was not being passed to the editing dialog.

Problem 2: when creating scoped external functions, the name of the
library was not being fully qualified. Scoped function libraries need
to be specified with a full path <bucket>/<scope>/<library name>.

Change-Id: I4a6848a40ef2418ae5d604352f34afa565f954a9
Reviewed-on: https://review.couchbase.org/c/query-ui/+/172084
Tested-by: Eben Haber <eben@couchbase.com>
Reviewed-by: Eben Haber <eben@couchbase.com>