MB-51528 - Show query workbench for user restricted to single scope
Users with query select on a single scope or collection were not
permitted to see the query workbench in 7.1. This was permitted in
7.0.3, however, and the query workbench did provide some limited
functionality.

The quick fix is to main.js, where we can change the ngShow to only
require docs.read, index.all, or meta.read.

However, a further fix is required to make the schema sidebar work
properly for the limited permission user. This is because the
scopes/collections REST API will not work for a user who only has
query select role on a single scope or collection. However, such a
user can still get scopes and collections via the query service:
select * from system:keyspaces (which only returns those keyspaces to
which the user is permitted to query).

Thus, the code in qw.query.workbench.service which formerly used the
ns_server REST API to get scopes and collections for buckets needed to
be changed to use a n1ql query instead.

Change-Id: Ia18f6cae81cf2311ebfad0c9786e1bee50404940
Reviewed-on: https://review.couchbase.org/c/query-ui/+/172631
Well-Formed: Restriction Checker
Tested-by: Eben Haber <eben@couchbase.com>
Reviewed-by: Pavel Blagodov <stochmail@gmail.com>