MB-51583 - Documents UI does not detect indexes for _default collecton
When we started retrieving index information from the query service
(when possible, for users with limited permissions not allowed to use
the regular REST APIs), I neglected to account for how indexes on
_default collections are described differently in the query
service. As with keyspace information, _default collections have an
undefined "bucket_id" field, instead putting the bucket name in the
"keyspace_id" (which is otherwise now used for collection names.

MB-51596: This also includes a fix for an exception thrown if the query to get
index information returns an error (usually due to the query service
not being ready).

Change-Id: Ia0f5936e045a2fe2afe0efea21e7059363f2fba9
Reviewed-on: https://review.couchbase.org/c/query-ui/+/172843
Reviewed-by: Pavel Blagodov <stochmail@gmail.com>
Well-Formed: Restriction Checker
Tested-by: Eben Haber <eben@couchbase.com>