MB-51579 - Doc Editor UI race condition when metadata not ready
There was a race condition with the Doc Editor trying to retrieve
documents when the bucket/scope/collection metadata was not yet
retrieved. A related problem occurred when the doc editor was launched
with no collection specified (as when the user clicks the "Documents"
link for a scope).

Change-Id: Ic1ab9ea7ba2377107a226f0fac90c89ab50d7f6b
Reviewed-on: https://review.couchbase.org/c/query-ui/+/172856
Well-Formed: Restriction Checker
Tested-by: Eben Haber <eben@couchbase.com>
Reviewed-by: Pavel Blagodov <stochmail@gmail.com>