MB-52041 - insufficient permissions to view doc editor on mixed cluster
In CB 7.0 and above, the doc editor requires permissions to retrieve
scope and collection names, to provide a menu for selecting
collections.

In mixed clusters 6.6 & 7.0, there are no collections, so such
permission should not be required.

This change adjusts permission required to see the doc editor.

This change also adjusts required permissions to see the query
workbench, the prior permissions were too restrictive.

Similarly, the permissions to use Query Import relied on
data.docs!write, which doesn't exist. Instead we needed to use
data.docs!upsert.

Change-Id: I8f56df9fd5f448242abd1584700d1ad064cc83a0
Reviewed-on: https://review.couchbase.org/c/query-ui/+/174709
Tested-by: Eben Haber <eben@couchbase.com>
Reviewed-by: Eben Haber <eben@couchbase.com>