MB-52550 - don't specify primary index when retrieving docs for doc editor
This is a forward merge from neo.

The document editor creates queries to page through documents in a
given collection. We want the documents in order of document ID, so to
do reliable paging we need to ask for documents ordered by doc ID. The
best way to get doc IDs in order is with a primary index (if one
exists), so in an attempt to improve doc editor queries, code was
added to explicitly name the primary index by adding a USE INDEX
clause to the query.

There were, however two problems with this.

1) The bug reported the doc editor mis-naming the primary index on the
statistics bucket, because the doc editor code assumed that the
primary index was named #primary, but the statistics bucket uses a
different name. In most cases the index in named #primary, but not
always.

2) A user noted in the Forums that they were using the doc ID with a
WHERE clause. They had a highly selective secondary index that matched
the WHERE clause, and in that case it is much more efficient to use
the secondary index to retrieve the doc IDs and sort them, versus
using the primary index, then filtering by the WHERE clause.

In sum, the addition of the USE INDEX clause was misguided, and it
should be removed. Instead I should just let the query optimizer do
what it does best.

Change-Id: I544ce71f6b0392d38264abd0a10b114044f5305a
Reviewed-on: https://review.couchbase.org/c/query-ui/+/176533
Well-Formed: Restriction Checker
Tested-by: Eben Haber <eben@couchbase.com>
Reviewed-by: Eben Haber <eben@couchbase.com>
Reviewed-on: https://review.couchbase.org/c/query-ui/+/176544