MB-52550 - Document page query correction
In 7.1.0 I tried to improve document page queries by forcing use of
the primary index when available. This improved performance in some
cases by eliminating the need for sorting the result set, but reduced
performance in the case of WHERE clauses that matched a highly
selective index.

The initial commit for this issue removed the "use index `#primary`"
from the query, but it neglected to ensure that the results are always
sorted by document ID (necessary to enuser that the LIMIT/OFFSET
batches don't overlap with each other). This change returns the "ORDER
BY" statement that was always present in 7.0.

Change-Id: Icbd15331c1abc636192041783b81cb3ddec1bd8f
Reviewed-on: https://review.couchbase.org/c/query-ui/+/176746
Well-Formed: Restriction Checker
Tested-by: Eben Haber <eben@couchbase.com>
Reviewed-by: Eben Haber <eben@couchbase.com>