MB-54232 - workbench display issues with array containing 'null' values. MB-54233 - workbench display issues with subobjects with numeric key.
Both these bugs came from CBSE reports, and both are related to
tabular representations of JSON documents (the read-only datatable
used in query results, and the spreadsheet-like tabular editor used in
the document viewer). While they describe two distinct bugs, the
underlying causes are similar: JSON corner cases.

Coming from a database background, I usually think of JSON documents
as tuples of values. JSON is much richer, however, as valid documents
also include arrays, primitive types, and even the value null. Both
the datatable and tabular editor try to show arrays of documents as
tables, with nested arrays as subtables.

Both bugs related to an optimization for dealing with nested arrays of
objects. The first occurred with arrays containing some objects, and
some null values. The code was checking for primitive values (numbers,
strings, and booleans) but not nulls, so it was trying to treat the
null value as an object. The second bug was related, involving
generated Angular code referencing the path to the value to be
displayed. E.g.,

   a.b.c.d

In this case, it was using the field name of a sub-object directly,
instead of referring to it with bracket notation. E.g.,

   a['b']['c']['d']

These two are equivalent in most cases, but if the subobject has a
name which is not a valid identifier (such as "12"), then only the
second notation is valid.

The change includes fixes for the two above bugs, but also includes
fixes for several un-reported issues with the same underlying causes:
incorrect handling of null values and heterogeneous arrays.

Change-Id: I923798d2dc3dd5e8f10c55821d11e030e5bcb64d
Reviewed-on: https://review.couchbase.org/c/query-ui/+/182419
Tested-by: Eben Haber <eben@couchbase.com>
Reviewed-by: Eben Haber <eben@couchbase.com>
Reviewed-on: https://review.couchbase.org/c/query-ui/+/182596
Well-Formed: Restriction Checker