MB-40918 - Move browser import toward parity with cbimport
One big limitation of the browser-based import in the lack of custom
document ID generation expressions. This change includes the first
step in that direction: a parser for ID expressions so the code can
understand what key pattern the user is describing.

The code does not yet use this, that will be the next step.

Change-Id: Iddc4897c995c2e490f43156ed925e0d63be03975
Reviewed-on: https://review.couchbase.org/c/query-ui/+/183392
Tested-by: Eben Haber <eben@couchbase.com>
Reviewed-by: Eben Haber <eben@couchbase.com>