MB-52940 - UI not retrieving documents on documents page
MB-54271 - UI sometimes doesn't show docs

The problem here occurred when navigating to the documents page. If
navigating from the list of scopes, the parameter wouldn't specify a
collection to open, so the code needed to default to the first
one. The code detecting whether default scopes were available was not
always working right, nor was it correctly making the selection when
it needed to.

Change-Id: I8a5abad1e7fe0be6baa9b1dcc69a9b455651fd0e
Reviewed-on: https://review.couchbase.org/c/query-ui/+/183825
Tested-by: Eben Haber <eben@couchbase.com>
Reviewed-by: Eben Haber <eben@couchbase.com>