MB-56014 - Display problems with Analytics query plan viewer
The most obvious problem here was that the list of fields at the top
of the Query Plan was expanding horizontally, making everything too
big for the window.

On closer examination, I noticed that the top was also showing an
empty list of "Buckets", even though the code was supposed to be
showing a list of "Datasets" (now called "Collections"). So there was
more than one problem to fix.

The first problem was fixed by adding several CSS styles to the
summary areas above the query plan. These styles hid any overflowing
content, made sure that the elements never got bigger than 25% of the
available space, and added an ellipsis if truncating the text. If the
data was truncated, however, it would be nice to show all the
information somewhere, so the list of all fields (or indexes or
datasets) is shown if the user hovers their mouse over the relevant
area.

The second problem was caused by the migration from AngularJS to
Angular 8. In the new version, the component was copying the query
plan data each time it changed. However it was only working correctly
for Query service plans, it was ignoring fields relevant to Analytics
plans (the 'mode' field, as well as the list of datasets referred to
in the plan).

Change-Id: I4651b836407d9a2a71e941b8db03e19f6bdbb409
Reviewed-on: https://review.couchbase.org/c/query-ui/+/188410
Well-Formed: Restriction Checker
Tested-by: Eben Haber <eben@couchbase.com>
Reviewed-by: Till Westmann <till@couchbase.com>