MB-58234 - JSON document causes UI docs panel to hang
As noted in the original bug report (see linked CBSE), the fix for
MB-54232 caused a regression in the documents view for very complex
JSON documents in the form of arrays.

The on-prem document viewer has two modes: field-editing mode
(spreadsheet like, where individual cells can be edited), and regular
mode, where up to 200 characters of the JSON are shown, and the user
must click the document ID to edit the document.

Field editing mode is generally not permitted when the document is
extremely complex. The browser gets stuck if you ask it to render
thousands of text fields.

The regression in MB-54232 meant that complex non-object documents
were not being checked for complexity. I.e., if a JSON document is a
very complex array instead of an object, the UI was trying to render a
field editor for thousands of fields, instead of just showing a
message saying that field editing is not possible. There are two parts
to the fix:

1) Around line 787: the getColumnArea() function determines how many
rows/columns a javascript object/array requires in field editing
mode. However, the 'count' of items was not being checked for objects
inside arrays.

2) Around line 636: this is the code that generates HTML for top-level
non-objects. If the array is too complex, an icon and error message is
shown, just as the same is shown later for objects that are too
complex.

Change-Id: I927ae22e6bfb03c5ab1d83043a7c0ce7e3e9ade2
Reviewed-on: https://review.couchbase.org/c/query-ui/+/195233
Tested-by: Eben Haber <eben@couchbase.com>
Well-Formed: Restriction Checker
Reviewed-by: Pavel Blagodov <stochmail@gmail.com>